move constraint up
