+++ squeal-postgresql/src/Squeal/PostgreSQL.hs
@@ -195,11 +195,11 @@ module Squeal.PostgreSQL (module X, RenderSQL(..), printSQL) where
 import Squeal.PostgreSQL.Alias as X
 import Squeal.PostgreSQL.Binary as X
 import Squeal.PostgreSQL.Definition as X
+import Squeal.PostgreSQL.Definition.Constraint as X
 import Squeal.PostgreSQL.Definition.Function as X
 import Squeal.PostgreSQL.Definition.Index as X
 import Squeal.PostgreSQL.Definition.Schema as X
 import Squeal.PostgreSQL.Definition.Table as X
-import Squeal.PostgreSQL.Definition.Table.Constraint as X
 import Squeal.PostgreSQL.Definition.Type as X
 import Squeal.PostgreSQL.Definition.View as X
 import Squeal.PostgreSQL.Expression as X
