+++ squeal-postgresql/src/Squeal/PostgreSQL/Definition/Table.hs
@@ -62,7 +62,7 @@ import qualified GHC.Generics as GHC
 
 import Squeal.PostgreSQL.Alias
 import Squeal.PostgreSQL.Definition
-import Squeal.PostgreSQL.Definition.Table.Constraint
+import Squeal.PostgreSQL.Definition.Constraint
 import Squeal.PostgreSQL.Expression
 import Squeal.PostgreSQL.Expression.Type
 import Squeal.PostgreSQL.List
