+++ squeal-postgresql/src/Squeal/PostgreSQL/Migration.hs
@@ -173,8 +173,8 @@ import qualified GHC.Generics as GHC
 import Squeal.PostgreSQL.Alias
 import Squeal.PostgreSQL.Binary
 import Squeal.PostgreSQL.Definition
+import Squeal.PostgreSQL.Definition.Constraint
 import Squeal.PostgreSQL.Definition.Table
-import Squeal.PostgreSQL.Definition.Table.Constraint
 import Squeal.PostgreSQL.Expression.Comparison
 import Squeal.PostgreSQL.Expression.Parameter
 import Squeal.PostgreSQL.Expression.Time
